deserialzation successful, recipe loaded

+++ Unity/Chef's Table/Assets/Scripts/Ingredient.cs
@@ -5,11 +5,11 @@ using UnityEngine;
 public class Ingredient
 {
     public string name;
-    public int quantity;
+    public double quantity;
     public string unit;
     private Vector3 postion;
 
-    public Ingredient(string name, int quantity, string unit)
+    public Ingredient(string name, double quantity, string unit)
     {
         this.name = name;
         this.quantity = quantity;
